backend/requirements.txt
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
binaryornot==0.4.4
boto3==1.35.68
botocore==1.35.90
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
cookiecutter==2.6.0
coverage==7.6.8
decorator==5.1.1
dj-database-url==2.3.0
Django==5.1.3
django-configurations==2.5.1
django-cors-headers==4.6.0
django-filter==24.3
django-model-utils==5.0.0
django-storages==1.14.4
django-unique-upload==0.2.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
docker==7.1.0
exceptiongroup==1.2.2
executing==2.1.0
factory_boy==3.3.1
Faker==33.1.0
flake8==7.1.1
ghp-import==2.1.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.5
jmespath==1.0.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mergedeep==1.3.4
mkdocs==1.6.1
mkdocs-get-deps==0.2.0
mock==5.1.0
newrelic==10.3.0
packaging==24.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
pyyaml_env_tag==0.1
requests==2.32.3
rich==13.9.4
s3transfer==0.10.4
six==1.17.0
sqlparse==0.5.3
stack-data==0.6.3
text-unidecode==1.3
tomli==2.2.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.3.0
watchdog==6.0.0
wcwidth==0.2.13
